Moved Repository layer to Chirp.Infrastructure instead of Core.
Mostly just means better organization of the project and namespaces. This also meant a little tweaking of where the tests are.
