+++ Chirp.sln
@@ -17,8 +17,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{AAF0C3E6-2
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Chirp.Core", "src\Chirp.Core\Chirp.Core.csproj", "{046F33B0-9573-4AFA-9A68-9B16344F9AF7}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Chirp.CoreTest", "test\Chirp.CoreTest\Chirp.CoreTest.csproj", "{A55A0AD5-17E3-46E3-BB49-A75FA973B78D}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Chirp.Razor", "src\Chirp.Razor\Chirp.Razor.csproj", "{812F9670-264D-4115-A73D-98ED6EE3CF8B}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Chirp.RazorTest", "test\Chirp.RazorTest\Chirp.RazorTest.csproj", "{537F69BC-02AC-4A5B-88DB-B77A9DBBC644}"
@@ -40,12 +38,6 @@ Global
 		{046F33B0-9573-4AFA-9A68-9B16344F9AF7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{046F33B0-9573-4AFA-9A68-9B16344F9AF7}.Test|Any CPU.ActiveCfg = Test|Any CPU
 		{046F33B0-9573-4AFA-9A68-9B16344F9AF7}.Test|Any CPU.Build.0 = Test|Any CPU
-		{A55A0AD5-17E3-46E3-BB49-A75FA973B78D}.Debug|Any CPU.ActiveCfg = Test|Any CPU
-		{A55A0AD5-17E3-46E3-BB49-A75FA973B78D}.Debug|Any CPU.Build.0 = Test|Any CPU
-		{A55A0AD5-17E3-46E3-BB49-A75FA973B78D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A55A0AD5-17E3-46E3-BB49-A75FA973B78D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A55A0AD5-17E3-46E3-BB49-A75FA973B78D}.Test|Any CPU.ActiveCfg = Test|Any CPU
-		{A55A0AD5-17E3-46E3-BB49-A75FA973B78D}.Test|Any CPU.Build.0 = Test|Any CPU
 		{812F9670-264D-4115-A73D-98ED6EE3CF8B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{812F9670-264D-4115-A73D-98ED6EE3CF8B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{812F9670-264D-4115-A73D-98ED6EE3CF8B}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -76,7 +68,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{046F33B0-9573-4AFA-9A68-9B16344F9AF7} = {8F8A6A34-59A8-446F-A994-A98F574F9A3E}
-		{A55A0AD5-17E3-46E3-BB49-A75FA973B78D} = {AAF0C3E6-2C01-4C84-9E9D-77812EF704AD}
 		{812F9670-264D-4115-A73D-98ED6EE3CF8B} = {8F8A6A34-59A8-446F-A994-A98F574F9A3E}
 		{537F69BC-02AC-4A5B-88DB-B77A9DBBC644} = {AAF0C3E6-2C01-4C84-9E9D-77812EF704AD}
 		{43E5398F-767D-4475-A8A4-FF51E2A8C06B} = {8F8A6A34-59A8-446F-A994-A98F574F9A3E}
