+++ src/Chirp.Infrastructure/Chirp.Infrastructure.csproj
@@ -11,4 +11,9 @@
         <ProjectReference Include="../Chirp.Core/Chirp.Core.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+        <Folder Include="Services\" />
+        <Folder Include="Repositories\"/>
+    </ItemGroup>
+
 </Project>
