+++ src/Chirp.Infrastructure/Services/AuthorService/AuthorViewModel.cs
@@ -1,3 +1,3 @@
-﻿namespace Chirp.Infrastructure.AuthorService;
+﻿namespace Chirp.Infrastructure.Services.AuthorService;
 
 public record AuthorViewModel(string Id, string Name, string Email);
