+++ src/Chirp.Infrastructure/Services/CheepService/CheepViewModel.cs
@@ -1,3 +1,3 @@
-﻿namespace Chirp.Infrastructure.CheepService;
+﻿namespace Chirp.Infrastructure.Services.CheepService;
 
 public record CheepViewModel(string Author, string Message, string Timestamp);
