+++ src/Chirp.Razor/Pages/Public.cshtml.cs
@@ -1,8 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-using Chirp.Core.Entities;
-using Chirp.Core.DataTransferObject;
-using Chirp.Infrastructure.CheepService;
+using Chirp.Infrastructure.Services.CheepService;
 
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
