+++ src/Chirp.Razor/Pages/UserTimeline.cshtml.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-using Chirp.Infrastructure.CheepService;
+using Chirp.Infrastructure.Services.CheepService;
 
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
