+++ src/Chirp.Razor/Program.cs
@@ -1,13 +1,10 @@
 using Chirp.Core;
 using Chirp.Core.Data;
-using Chirp.Core.Repositories;
-using Chirp.Infrastructure.CheepService;
 
 using Microsoft.EntityFrameworkCore;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.Authentication.OAuth;
 using Chirp.Core.Entities;
+using Chirp.Infrastructure.Repositories;
+using Chirp.Infrastructure.Services.CheepService;
 
 namespace Chirp.Razor;
