+++ test/Chirp.InfrastructureTest/RepositoryTest/CheepRepositoryUnitTest.cs
@@ -2,9 +2,9 @@
 
 using Chirp.Core.DataTransferObject;
 using Chirp.Core.Models;
-using Chirp.Core.Repositories;
+using Chirp.Infrastructure.Repositories;
 
-namespace Chirp.CoreTest;
+namespace Chirp.InfrastructureTest.RepositoryTest;
 
 public class CheepRepositoryUnitTest : CoreRepositoryTester
 {
